moved starcat5_hpic_merger files
